agents/requirements.txt
livekit-agents[hedra]~=1.2
livekit-agents[images]~=1.2
python-dotenv~=1.0.0

